+++ examples/physics/demo_blur_tour.py
@@ -11,7 +11,7 @@ import torch
 
 import deepinv as dinv
 from deepinv.utils.plotting import plot
-from deepinv.utils.demo import load_example
+from deepinv.utils import load_example
 
 
 # %% Load test images