+++ examples/self-supervised-learning/demo_unsure.py
@@ -24,7 +24,7 @@ from torch.utils.data import DataLoader
 from torchvision import transforms, datasets
 
 import deepinv as dinv
-from deepinv.utils.demo import get_data_home
+from deepinv.utils import get_data_home
 
 # %%
 # Setup paths for data loading and results.